+++ src/js/util/index.js
@@ -12,6 +12,7 @@ export * from './lang';
 export * from './mouse';
 export * from './options';
 export * from './player';
+export * from './position';
 export * from './promise';
 export * from './intersection';
 export * from './selector';